remote items: fix item unobtainable bug
if someone checks the same location in a different world and sends that item to this world, and this world has remote items on, the item same location in this world disappears.
the same bug is already fixed for archipelago 0.3.3 client users by a fix in SNIClient. this commit should fix the bug for users of *earlier clients* when rolling on AP 0.3.4+ server